+++ src/app/shared/empty-state/empty-state.component.ts
@@ -1,0 +1,14 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'app-empty-state',
+  templateUrl: './empty-state.component.html',
+  styleUrls: ['./empty-state.component.scss']
+})
+export class EmptyStateComponent {
+  @Input() image: string;
+  @Input() title: string;
+  @Input() description: string;
+
+  constructor() { }
+}