+++ src/app/views/loan-detail/detail-installments/detail-installments.component.spec.ts
@@ -2,10 +2,10 @@ import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { HttpClientModule } from '@angular/common/http';
 import { DetailInstallmentsComponent } from './detail-installments.component';
+import { SharedModule } from './../../../shared/shared.module';
 import { InstallmentsService } from './../../../services/installments.service';
 import { EventsService } from './../../../services/events.service';
 import { CommitsService } from './../../../services/commits.service';
-import { FormatAmountPipe } from './../../../pipes/format-amount.pipe';
 
 describe('DetailInstallmentsComponent', () => {
   let component: DetailInstallmentsComponent;
@@ -13,8 +13,8 @@ describe('DetailInstallmentsComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [ HttpClientModule ],
-      declarations: [ DetailInstallmentsComponent, FormatAmountPipe ],
+      imports: [ HttpClientModule, SharedModule ],
+      declarations: [ DetailInstallmentsComponent ],
       providers: [ InstallmentsService, EventsService, CommitsService ],
       schemas: [ CUSTOM_ELEMENTS_SCHEMA ]
     })