+++ src/app/views/loan-detail/loan-detail.component.ts
@@ -24,11 +24,11 @@ import { LoanTypeService } from './../../services/loan-type.service';
 import { PreviousRouteService } from '../../services/previousRoute.service';
 
 @Component({
-  selector: 'app-loan-detail2',
-  templateUrl: './loan-detail2.component.html',
-  styleUrls: ['./loan-detail2.component.scss']
+  selector: 'app-loan-detail',
+  templateUrl: './loan-detail.component.html',
+  styleUrls: ['./loan-detail.component.scss']
 })
-export class LoanDetail2Component implements OnInit, OnDestroy {
+export class LoanDetailComponent implements OnInit, OnDestroy {
   pageId = 'loan-detail';
   loan: Loan;
   identityName: string;
@@ -77,7 +77,12 @@ export class LoanDetail2Component implements OnInit, OnDestroy {
   punitory: string;
   paymentDate: string[] = [];
   paymentAverage: string;
-  timelineTooltip: string;
+  paymentAverageAmount: number;
+  timeline: {
+    tooltip: string;
+    icon: string;
+    iconType: 'material' | 'fontawesome' | 'fontello' | 'image';
+  };
 
   // Loan Oracle
   oracle: string;
@@ -358,9 +363,17 @@ export class LoanDetail2Component implements OnInit, OnDestroy {
           const today = Math.floor(new Date().getTime() / 1000);
           const expiresIn = Utils.formatDelta(this.loan.expiration - today);
           this.expiresIn = expiresIn;
-          this.timelineTooltip = '< > Requested';
+          this.timeline = {
+            tooltip: '< > Requested',
+            icon: 'code',
+            iconType: 'material'
+          };
         } else {
-          this.timelineTooltip = 'Expired';
+          this.timeline = {
+            tooltip: 'Expired',
+            icon: 'delete',
+            iconType: 'material'
+          };
         }
         break;
       case Status.Indebt:
@@ -411,8 +424,19 @@ export class LoanDetail2Component implements OnInit, OnDestroy {
           this.durationTooltip = `Next payment in ${ durationDynamic }`;
         }
 
-        this.timelineTooltip =
-          this.loan.status === Status.Paid ? 'Fully Paid' : 'Outstanding';
+        if (this.loan.status === Status.Paid) {
+          this.timeline = {
+            tooltip: 'Fully Paid',
+            icon: 'icon-verified-24px',
+            iconType: 'fontello'
+          };
+        } else {
+          this.timeline = {
+            tooltip: 'Outstanding',
+            icon: 'trending_up',
+            iconType: 'material'
+          };
+        }
         break;
 
       default:
@@ -480,6 +504,7 @@ export class LoanDetail2Component implements OnInit, OnDestroy {
     const diffToNowDays = (endDate - nowDate) / SECONDS_IN_DAY;
     const daysAverage = 100 - ((diffToNowDays * 100) / diffDays);
     this.paymentAverage = `${ daysAverage > MAX_AVERAGE ? MAX_AVERAGE : daysAverage }%`;
+    this.paymentAverageAmount = Math.round(daysAverage > MAX_AVERAGE ? MAX_AVERAGE : daysAverage);
   }
 
   private invalidActions() {